Final Project: Ready for Review
In this section of the work day I had to correct whatever aside errors that I made so that the page content is on the side of each other. When the changes were made, the page looked right; the biggest issues were the child of rules that I had. I would include smaller things like ol and ul to make it not readable. In the end, I worked on the section, article, aside, and outline IDs.